feat(editor): add constrained free rotation

+++ src/GeometryDashPlace.Web/Components/Editor/State/EditorObjectInstance.cs
@@ -5,7 +5,7 @@ public sealed class EditorObjectInstance
     public required string Type { get; set; }
     public int X { get; set; }
     public int Y { get; set; }
-    public int Rotation { get; set; }
+    public double Rotation { get; set; }
     public string? ColorTarget { get; set; }
     public int Red { get; set; } = 255;
     public int Green { get; set; } = 255;
